join game, start game end points added numelens to game entity create first round, select random question

+++ src/main/java/com/bhanu/psychiba/model/Round.java
@@ -23,7 +23,7 @@ public class Round extends Auditable {
     @NotNull
     @Getter
     @Setter
-    private Long roundNumber;
+    private int roundNumber;
 
     @ManyToOne
     @Getter
